Adds Refactoring of Keygen Class
